Delete orders: new ACL, and depends on config param
git-svn-id: https://dev.virtuemart.net/svn/virtuemart@9792 87cc427b-4a1e-0410-b729-b2260ec0f063

+++ branches/com_virtuemart.3.0.12.4/administrator/components/com_virtuemart/views/orders/view.html.php
@@ -257,7 +257,7 @@ class VirtuemartViewOrders extends VmViewAdmin {
 
 			JToolBarHelper::save('updatestatus', vmText::_('COM_VIRTUEMART_UPDATE_STATUS'));
 
-			if (vmAccess::manager('orders.delete')) {
+			if (vmAccess::manager('orders.delete') && !VmConfig::get('ordersAddOnly',false)) {
 				JToolBarHelper::spacer('80');
 				JToolBarHelper::deleteList();
 			}
